Refactor FungibleType to TokensKind for Token permissions
a variant of FungibleType is needed for TokenParticle, so will refactor for that purpose only
