+++ radixdlt-java/src/main/java/com/radixdlt/client/atommodel/tokens/FeeParticle.java
@@ -1,7 +1,6 @@
 package com.radixdlt.client.atommodel.tokens;
 
 import com.radixdlt.client.application.translate.tokens.TokenTypeReference;
-import com.radixdlt.client.atommodel.FungibleType;
 import com.radixdlt.client.core.atoms.particles.Particle;
 import com.radixdlt.client.core.atoms.particles.RadixResourceIdentifer;
 import com.radixdlt.client.core.crypto.ECPublicKey;
