+++ radixdlt-java/src/test/java/com/radixdlt/client/application/translate/tokenclasses/TokenTypesReducerTest.java
@@ -8,7 +8,7 @@ import org.junit.Test;
 import org.radix.utils.UInt256;
 
 import com.radixdlt.client.application.translate.tokenclasses.TokenState.TokenSupplyType;
-import com.radixdlt.client.atommodel.FungibleType;
+import com.radixdlt.client.atommodel.tokens.TokensKind;
 import com.radixdlt.client.application.translate.tokens.TokenTypeReference;
 import com.radixdlt.client.atommodel.tokens.TokenParticle;
 import com.radixdlt.client.atommodel.tokens.TokenPermission;
@@ -28,7 +28,7 @@ public class TokenTypesReducerTest {
 		when(tokenParticle.getSymbol()).thenReturn("ISO");
 		when(tokenParticle.getDescription()).thenReturn("Desc");
 		when(tokenParticle.getGranularity()).thenReturn(UInt256.ONE);
-		when(tokenParticle.getTokenPermissions()).thenReturn(Collections.singletonMap(FungibleType.MINTED, TokenPermission.SAME_ATOM_ONLY));
+		when(tokenParticle.getTokenPermissions()).thenReturn(Collections.singletonMap(TokensKind.MINTED, TokenPermission.SAME_ATOM_ONLY));
 
 		TokenTypesReducer tokenTypesReducer = new TokenTypesReducer();
 		TokenTypesState state = tokenTypesReducer.reduce(TokenTypesState.init(), SpunParticle.up(tokenParticle));
@@ -48,7 +48,7 @@ public class TokenTypesReducerTest {
 		when(tokenParticle.getSymbol()).thenReturn("ISO");
 		when(tokenParticle.getDescription()).thenReturn("Desc");
 		when(tokenParticle.getGranularity()).thenReturn(UInt256.ONE);
-		when(tokenParticle.getTokenPermissions()).thenReturn(Collections.singletonMap(FungibleType.MINTED, TokenPermission.TOKEN_OWNER_ONLY));
+		when(tokenParticle.getTokenPermissions()).thenReturn(Collections.singletonMap(TokensKind.MINTED, TokenPermission.TOKEN_OWNER_ONLY));
 
 		MintedTokensParticle minted = mock(MintedTokensParticle.class);
 		when(minted.getAmount()).thenReturn(hundred);
